backup_before_refactor/pyproject.toml
[project]
name = "organoid-embedding-experiments"
version = "0.1.0"
description = "Research codebase for comparing embedding methods on liver organoid oxygen time series data"
authors = [
    {name = "Research Team", email = "research@example.com"},
]
dependencies = [
    # Core dependencies
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    # Supabase
    "supabase>=2.4.0",
    "python-dotenv>=1.0.0",
    # Time series libraries
    "tslearn>=0.6.0",
    "tsfresh>=0.20.0",
    "pycatch22>=0.4.0",
    "pyts>=0.12.0",
    "stumpy>=1.11.0",
    # Deep learning
    "torch>=2.0.0",
    "tensorflow>=2.15.0",
    # Visualization
    "plotly>=5.15.0",
    "umap-learn>=0.5.0",
    # Utilities
    "tqdm>=4.65.0",
    "pyyaml>=6.0.0",
    "joblib>=1.3.0",
    "hydra-core>=1.3.0",
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["data", "embeddings", "evaluation", "experiments"]

[tool.uv]
dev-dependencies = [
    "jupyter>=1.0.0",
    "pytest>=7.0.0",
]
